New build system implemented and updated readme
This new build system accomplishes the following:
    - The object files are abstracted away, omitting the need for a
      'make clean'.
    - A debug build is automatically appended to the build process.
    - Specified language standard.  c99 is faster and generally more
      trusted over c23 or c2x.
    - Flags for both the prod and debug build are added.  Modifying
      the compiler flags is also really easy.
    - Enables multiple builddirs.

I forked this repo, tried to './main init .' and was met with a
segfault.  I'm sure it's a temporary bug and a quick fix but I think
for a repo this size, having a debug build ready to go out of the box
is massive plus.

Let me know if you think this build system could use a little work.
Off the top of my head, ensuring that this builds on windows machines
wouldn't hurt.

As for the readme, I tried to follow the same style for the appended
build section.  I'm sure it wouldn't take too much modification if it
isn't up to the project standards.